pyproject.toml
[project]
name = "fastapi-audio-serv"
version = "0.1.0"
description = ""
authors = [
    {name = "Pertsev Sergey",email = "pertsev.sergey59@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "aiohttp (>=3.11.14,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic[email] (>=2.11.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false